clean up hypothesis modules and functions
